refactor: /utils/helper로 변경 (#43)
